requirements.txt
Django==1.6.1
Markdown==2.3.1
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-bootstrapped==1.6.3
django-fields==0.2.2
django-filter==0.7
django-ios-notifications==0.1.5
django-registration==1.0
django-suit==0.2.5
django-toolbelt==0.0.1
djangorestframework==2.3.10
gunicorn==18.0
psycopg2==2.5.1
pyOpenSSL==0.13.1
pycrypto==2.6.1
static==0.4
wsgiref==0.1.2
